C++ API
Former-commit-id: 70b5f77d758f66aeb48a8eb11ebce27df9f251fc

+++ api/c++/cub.cc
@@ -23,7 +23,7 @@ CUB::~CUB()
 //------------------------------------------------------------
 void CUB::headin()
 {
-    off_t nn[SF_MAX_DIM];
+    int nn[SF_MAX_DIM];
     char key[3];
 
     nd = sf_filedims(file_,nn);
